app/requirements.txt
fastapi
uvicorn
langchain
groq
transformers
torch
pandas
qdrant-client
sqlalchemy
psycopg2-binary
boto3
fpdf
python-dotenv
